Fixes #1871 

This PR's branch is based on top of the branch of PR #1898, it only makes sense to review this PR once #1898 is merged.

This adds an optimization which is specific for queries that query for the absence of a meta tag of which all underlying expressions are using the `=` operator. This optimization currently already exists for queries which query for the presence of a meta tag using `=` and of which all underlying expressions are also using the `=` operator, in this PR that existing optimization gets extended to also work for filter expressions which check for the absence of a meta tag.

I also added a benchmark to test for that specific case, and It shows significant improvements with/without this change:

```
benchmark                                old ns/op        new ns/op     delta
BenchmarkFilterByAbsenceOfMetaTag-12     207231916200     253464875     -99.88%

benchmark                                old allocs     new allocs     delta
BenchmarkFilterByAbsenceOfMetaTag-12     407059         232777         -42.81%

benchmark                                old bytes     new bytes     delta
BenchmarkFilterByAbsenceOfMetaTag-12     37499976      34272224      -8.61%
```